requirements.txt
fastapi
uvicorn
python-multipart
openpyxl
google-generativeai
requests
python-dotenv
jinja2
